pyproject.toml
[project]
name = "markapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "fastapi[standard]>=0.134.0",
    "openai>=2.24.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
